refactor: group module directories by its similarites
