+++ src/index.js
@@ -1,6 +1,6 @@
 import createHeaderPage from './components/header-page.js';
-import VendingMachine from './machine.js';
-import { checkStorage } from './storage/util.js';
+import VendingMachine from './machine/machine.js';
+import { checkStorage } from './utils/storage.js';
 import createEventListeners from './components/eventListeners.js';
 
 createHeaderPage();
