+++ src/machine/machine.js
@@ -1,8 +1,8 @@
-import { ADD_TAB_ID, ADD_TAB_CLASS, MANAGE_TAB_ID, COIN_VALUE, COIN_TABLE_ID, PURCHASE_TAB_ID, PURCHASE_TAB_CLASS, PURCHASE_TAB_DATASET, ERROR_MESSAGE } from './constants.js';
+import { ADD_TAB_ID, ADD_TAB_CLASS, MANAGE_TAB_ID, COIN_VALUE, COIN_TABLE_ID, PURCHASE_TAB_ID, PURCHASE_TAB_CLASS, PURCHASE_TAB_DATASET, ERROR_MESSAGE } from '../constants.js';
 import Product from './product.js'
-import { elementCreatorWithClass } from './dom/util.js';
+import { elementCreatorWithClass } from '../utils/dom.js';
 import * as validator from './validatior.js';
-import { saveStorage } from './storage/util.js';
+import { saveStorage } from '../utils/storage.js';
 
 export default class VendingMachine {
     constructor(){
