+++ src/machine/product.js
